requirements.txt
# Core Framework
langgraph==0.2.75
langchain==0.2.17
langchain-core==0.2.43
langchain-openai==0.1.25

# Web Framework
fastapi==0.121.2
uvicorn==0.38.0
pydantic==2.12.3

# Search APIs
google-search-results==2.4.2

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
requests==2.32.5
httpx==0.28.1
jinja2==3.1.6

# Image Processing
pillow==10.3.0

# Optional: Local LLM
# ollama==0.1.6

# Development
pytest==8.3.5
pytest-asyncio==0.23.8
